All files: Format string fixups
- Use unsigned integers instead of signed integers where it
  makes sense to do so

- Use %u in format strings for unsigned integers; %d is for
  signed integers

- Kill some atoi() on user-provided data and parse it properly
  as an unsigned integer with error detection instead

  Protocol modules may still use it, as IRCds are assumed not
  to send absolute nonsense our way

  Using it on md->value for some `struct metadata *md` in the
  private: namespace is also supposed to be okay

  Many paths using atoi() that don't have the capability for
  returning errors are also left as-is. Consider this work
  a preliminary stab at the issue.

- Move some widely-shared numeric and string constants to a
  new header file and make the numeric constants unsigned

  Add the header to po/POTFILES.in so that its strings can be
  translated

+++ include/atheme/account.h
@@ -76,8 +76,8 @@ struct myuser
 	struct soper *          soper;
 	unsigned int            flags;
 	mowgli_list_t           memos;                  // store memos
-	unsigned short          memoct_new;
-	unsigned short          memo_ratelimit_num;     // memos sent recently
+	unsigned int            memoct_new;
+	unsigned int            memo_ratelimit_num;     // memos sent recently
 	time_t                  memo_ratelimit_time;    // last time a memo was sent
 	mowgli_list_t           memo_ignores;
 	mowgli_list_t           access_list;
